+++ sdk/formrecognizer/azure-ai-formrecognizer/src/samples/java/com/azure/ai/formrecognizer/FormRecognizerAsyncClientJavaDocCodeSnippets.java
@@ -114,7 +114,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
         // BEGIN: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeCustomForms#Flux-string-long-FormContentType
         File sourceFile = new File("{file_source_url}");
         String modelId = "{model_id}";
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
 
         formRecognizerAsyncClient.beginRecognizeCustomForms(buffer, modelId, sourceFile.length(),
@@ -144,7 +144,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
         File sourceFile = new File("{file_source_url}");
         String modelId = "{model_id}";
         boolean includeTextDetails = true;
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
 
         formRecognizerAsyncClient.beginRecognizeCustomForms(buffer, modelId, sourceFile.length(),
@@ -218,7 +218,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
     public void beginRecognizeContent() throws IOException {
         // BEGIN: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeContent#Flux-long-FormContentType
         File sourceFile = new File("{file_source_url}");
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
 
         formRecognizerAsyncClient.beginRecognizeContent(buffer, sourceFile.length(), FormContentType.APPLICATION_PDF)
@@ -245,7 +245,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
     public void beginRecognizeContentWithOptions() throws IOException {
         // BEGIN: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeContent#Flux-FormContentType-long-Duration
         File sourceFile = new File("{file_source_url}");
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
 
         formRecognizerAsyncClient.beginRecognizeContent(buffer, FormContentType.APPLICATION_PDF, sourceFile.length(),
@@ -333,7 +333,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
     public void beginRecognizeReceipts() throws IOException {
         // BEGIN: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeReceipts#Flux-long-FormContentType
         File sourceFile = new File("{file_source_url}");
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
         formRecognizerAsyncClient.beginRecognizeReceipts(buffer, sourceFile.length(), FormContentType.IMAGE_JPEG)
             .subscribe(recognizePollingOperation -> {
@@ -369,7 +369,7 @@ public class FormRecognizerAsyncClientJavaDocCodeSnippets {
         // BEGIN: com.azure.ai.formrecognizer.FormRecognizerAsyncClient.beginRecognizeReceipts#Flux-long-FormContentType-boolean-Duration
         File sourceFile = new File("{file_source_url}");
         boolean includeTextDetails = true;
-        Flux<ByteBuffer> buffer = Utility.convertStreamToByteBuffer(
+        Flux<ByteBuffer> buffer = Utility.toFluxByteBuffer(
             new ByteArrayInputStream(Files.readAllBytes(sourceFile.toPath())));
         formRecognizerAsyncClient.beginRecognizeReceipts(buffer, sourceFile.length(), FormContentType.IMAGE_JPEG,
             includeTextDetails, Duration.ofSeconds(5)).subscribe(recognizePollingOperation -> {