+++ sdk/formrecognizer/azure-ai-formrecognizer/src/test/java/com/azure/ai/formrecognizer/FormRecognizerAsyncClientTest.java
@@ -3,7 +3,6 @@
 
 package com.azure.ai.formrecognizer;
 
-import com.azure.ai.formrecognizer.implementation.Utility;
 import com.azure.ai.formrecognizer.models.CustomFormModel;
 import com.azure.ai.formrecognizer.models.ErrorResponseException;
 import com.azure.ai.formrecognizer.models.FormContentType;
@@ -25,12 +24,14 @@ import static com.azure.ai.formrecognizer.TestUtils.INVALID_SOURCE_URL_ERROR;
 import static com.azure.ai.formrecognizer.TestUtils.INVALID_URL;
 import static com.azure.ai.formrecognizer.TestUtils.LAYOUT_FILE_LENGTH;
 import static com.azure.ai.formrecognizer.TestUtils.RECEIPT_FILE_LENGTH;
+import static com.azure.ai.formrecognizer.TestUtils.RECEIPT_LOCAL_URL;
 import static com.azure.ai.formrecognizer.TestUtils.getExpectedFormPages;
 import static com.azure.ai.formrecognizer.TestUtils.getExpectedReceipts;
 import static com.azure.ai.formrecognizer.TestUtils.getExpectedRecognizedForms;
 import static com.azure.ai.formrecognizer.TestUtils.getExpectedRecognizedLabeledForms;
 import static com.azure.ai.formrecognizer.TestUtils.getExpectedUSReceipt;
-import static com.azure.ai.formrecognizer.TestUtils.getFileBufferData;
+import static com.azure.ai.formrecognizer.TestUtils.getReplayableBufferData;
+import static com.azure.ai.formrecognizer.implementation.Utility.toFluxByteBuffer;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
@@ -90,7 +91,7 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
     void recognizeReceiptData() {
         receiptDataRunner((data) -> {
             SyncPoller<OperationResult, IterableStream<RecognizedReceipt>> syncPoller =
-                client.beginRecognizeReceipts(getFileBufferData(data), RECEIPT_FILE_LENGTH,
+                client.beginRecognizeReceipts(toFluxByteBuffer(data), RECEIPT_FILE_LENGTH,
                     FormContentType.IMAGE_JPEG, false,
                     null).getSyncPoller();
             syncPoller.waitForCompletion();
@@ -105,7 +106,7 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
     void recognizeReceiptDataTextDetails() {
         receiptDataRunnerTextDetails((data, includeTextDetails) -> {
             SyncPoller<OperationResult, IterableStream<RecognizedReceipt>> syncPoller
-                = client.beginRecognizeReceipts(getFileBufferData(data), RECEIPT_FILE_LENGTH,
+                = client.beginRecognizeReceipts(toFluxByteBuffer(data), RECEIPT_FILE_LENGTH,
                 FormContentType.IMAGE_JPEG, includeTextDetails,
                 null).getSyncPoller();
             syncPoller.waitForCompletion();
@@ -130,7 +131,7 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
     void recognizeReceiptAsUSReceipt() {
         receiptDataRunnerTextDetails((data, includeTextDetails) -> {
             SyncPoller<OperationResult, IterableStream<RecognizedReceipt>> syncPoller
-                = client.beginRecognizeReceipts(getFileBufferData(data), RECEIPT_FILE_LENGTH,
+                = client.beginRecognizeReceipts(toFluxByteBuffer(data), RECEIPT_FILE_LENGTH,
                 FormContentType.IMAGE_JPEG, includeTextDetails,
                 null).getSyncPoller();
             syncPoller.waitForCompletion();
@@ -140,34 +141,31 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
     }
 
     /**
+     * Verifies receipt data from a document using file data as source.
+     * And the content type is not given. The content will be auto detected.
+     */
+    @Test
+    void recognizeReceiptDataWithContentTypeAutoDetection() {
+        SyncPoller<OperationResult, IterableStream<RecognizedReceipt>> syncPoller =
+            client.beginRecognizeReceipts(getReplayableBufferData(RECEIPT_LOCAL_URL), RECEIPT_FILE_LENGTH, null,
+                false, null).getSyncPoller();
+        syncPoller.waitForCompletion();
+        validateReceiptResult(getExpectedReceipts(false), syncPoller.getFinalResult());
+    }
+
+    /**
      * Verifies layout data for a document using source as input stream data.
      */
     @Test
     void recognizeLayoutData() {
         layoutDataRunner((data) -> {
             SyncPoller<OperationResult, IterableStream<FormPage>> syncPoller
-                = client.beginRecognizeContent(getFileBufferData(data),
+                = client.beginRecognizeContent(toFluxByteBuffer(data),
                 FormContentType.IMAGE_JPEG, LAYOUT_FILE_LENGTH, null).getSyncPoller();
             syncPoller.waitForCompletion();
             validateLayoutResult(getExpectedFormPages(), syncPoller.getFinalResult());
         });
     }
-
-    // TODO: once PR https://github.com/Azure/azure-sdk-for-java/pull/10354/files is merged, record it
-//    /**
-//     * Verifies layout data for a document using source as input stream data.
-//     * And the content type is not given. The content will be auto detected.
-//     */
-//    @Test
-//    void extractLayoutValidSourceWithContentTypeAutoDetection() {
-//        layoutValidSourceUrlRunner((data) -> {
-//            SyncPoller<OperationResult, IterableStream<FormPage>> syncPoller
-//                = client.beginRecognizeContent(getFileBufferData(data),
-//                null, LAYOUT_FILE_LENGTH, null).getSyncPoller();
-//            syncPoller.waitForCompletion();
-//            validateLayoutResult(getExpectedFormPages(), syncPoller.getFinalResult());
-//        });
-//    }
 
     /**
      * Verifies layout data for a document using source as input stream data.
@@ -202,19 +200,9 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
             syncPoller.waitForCompletion();
             CustomFormModel createdModel = syncPoller.getFinalResult();
             StepVerifier.create(client.beginRecognizeCustomFormsFromUrl(INVALID_URL,
-                createdModel.getModelId())).verifyErrorSatisfies(throwable -> assertEquals(throwable.getMessage(), INVALID_SOURCE_URL_ERROR));
+                createdModel.getModelId())).verifyErrorSatisfies(throwable -> assertEquals(throwable.getMessage(),
+                INVALID_SOURCE_URL_ERROR));
         });
-    }
-
-    /**
-     * Verifies that an exception is thrown for invalid status model Id.
-     */
-    @Test
-    void extractCustomFormInValidSourceUrl() {
-        ErrorResponseException httpResponseException = assertThrows(
-            ErrorResponseException.class,
-            () -> client.beginRecognizeCustomFormsFromUrl(INVALID_URL, VALID_MODEL_ID).getSyncPoller().getFinalResult());
-        assertEquals(httpResponseException.getMessage(), (INVALID_SOURCE_URL_ERROR));
     }
 
     /**
@@ -229,7 +217,7 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
                 syncPoller.waitForCompletion();
 
                 SyncPoller<OperationResult, IterableStream<RecognizedForm>> syncPollers
-                    = client.beginRecognizeCustomForms(Utility.convertStreamToByteBuffer(data),
+                    = client.beginRecognizeCustomForms(toFluxByteBuffer(data),
                     syncPoller.getFinalResult().getModelId(),
                     CUSTOM_FORM_FILE_LENGTH, FormContentType.APPLICATION_PDF, true, null)
                     .getSyncPoller();
@@ -250,7 +238,7 @@ public class FormRecognizerAsyncClientTest extends FormRecognizerClientTestBase 
                 syncPoller.waitForCompletion();
 
                 SyncPoller<OperationResult, IterableStream<RecognizedForm>> syncPollers
-                    = client.beginRecognizeCustomForms(Utility.convertStreamToByteBuffer(data),
+                    = client.beginRecognizeCustomForms(toFluxByteBuffer(data),
                     syncPoller.getFinalResult().getModelId(),
                     CUSTOM_FORM_FILE_LENGTH, FormContentType.APPLICATION_PDF, false, null)
                     .getSyncPoller();